requirements.txt
certifi==2017.4.17
chardet==3.0.4
django-allauth==0.32.0
django-bower==5.2.0
django-compressor==2.1.1
django-extensions==1.7.9
django-sass-processor==0.5.4
libsass==0.13.2
oauthlib==2.0.2
python-openid==2.2.5
requests==2.18.1
requests-oauthlib==0.8.0
urllib3==1.21.1
